# api/consumer_api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenStudyBuilder Consumer API",
...
    "version": "0.1.150"
  },
  "paths": {
...
    "/v1/studies": {
      "get": {
        "tags": [
          "[V1] Studies"
        ],
        "summary": "Get Studies",
        "description": "Returns a paginated list of studies, sorted by the specified sort criteria and order.\n\nEach returned study contains a full list of corresponding study versions, sorted by version start date in descending order.\n\nReturned `version_number` value can be used in other endpoints to retrieve study entities (e.g. visits, activities, etc.)\nassociated with a specific study version.\n\nCodelist details can be retrieved from the `GET /v1/library/ct/codelists` endpoint.\n\nDetails related to the Data Supplier type can be retrieved from the `GET /v1/library/ct/codelist-terms?codelist_submission_value=DATA_SUPPLIER_TYPE` endpoint.",
        "operationId": "get_studies_v1_studies_get",
        "security": [
          {
            "OAuth2AuthorizationCodeBearer": []
          }
        ],
        "parameters": [
          {
            "name": "sort_by",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/SortByStudies",
              "default": "uid"
            }
          },
          {
            "name": "sort_order",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/SortOrder",
              "default": "asc"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "description": "Number of items to be returned per page. Must be between 1 and the maximum allowed page size.",
              "default": 10,
              "title": "Page Size"
            },
            "description": "Number of items to be returned per page. Must be between 1 and the maximum allowed page size."
          },
          {
            "name": "page_number",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000000000,
              "minimum": 1,
              "description": "Page number of the returned list of entities. Must be between 1 and the maximum allowed page number.",
              "default": 1,
              "title": "Page Number"
            },
            "description": "Page number of the returned list of entities. Must be between 1 and the maximum allowed page number."
          },
          {
            "name": "id",
            "in": "query",
            "required": false,
            "schema": {
              "description": "Filter by study ID (case-insensitive partial match), for example `XX1234-5678`.",
              "title": "Id",
              "type": "string"
            },
            "description": "Filter by study ID (case-insensitive partial match), for example `XX1234-5678`."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_Study_"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "[V1] Studies"
        ],
        "summary": "Create Study",
        "description": "Creates a new study definition in `DRAFT` status.\n\n- Only top-level studies can be created here. Study subparts are not supported\nby this endpoint.\n\n- At least one of `study_number` and `study_acronym` must be provided.\n\n- `study_number` and `study_acronym` must each be unique across all studies,\nnot just within the project. A collision is rejected with HTTP 409.\n\n- The study ID is derived by the server as `{project_id}-{study_number}`\nand cannot be supplied by the caller.\n\n<br/>\nValid values for payload fields:\n\n- **project_id**: Allowed values are retrieved from the `GET /v1/library/projects`\nendpoint (**id** field). The value is matched exactly and case-sensitively.",
        "operationId": "create_study_v1_studies_post",
        "security": [
          {
            "OAuth2AuthorizationCodeBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StudyCreateInput"
              },
              "examples": {
                "example": {
                  "summary": "Create a study",
                  "value": {
                    "study_number": "5678",
                    "study_acronym": "STUDY-ACRONYM",
                    "project_id": "AB1234",
                    "description": "Study description"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Study created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudyCreateResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict - a study with the same number or acronym already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "properties": {
          "time": {
            "type": "string",
            "format": "date-time",
            "title": "Time"
          },
          "path": {
            "type": "string",
            "title": "Path"
          },
          "method": {
            "type": "string",
            "title": "Method"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "details": {
            "items": {
              "$ref": "#/components/schemas/ValidationDetail"
            },
            "type": "array",
            "title": "Details"
          }
        },
        "type": "object",
        "required": [
          "time",
          "path",
          "method",
          "type",
          "message"
        ],
        "title": "ErrorResponse"
      },
...
      "PaginatedResponse_Study_": {
        "properties": {
          "self": {
            "type": "string",
            "title": "Self",
            "description": "Pagination link pointing to the current page"
          },
          "prev": {
            "type": "string",
            "title": "Prev",
            "description": "Pagination link pointing to the previous page"
          },
          "next": {
            "type": "string",
            "title": "Next",
            "description": "Pagination link pointing to the next page"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/Study"
            },
            "type": "array",
            "title": "Items",
            "description": "List of items"
          }
        },
        "type": "object",
        "required": [
          "self",
          "prev",
          "next",
          "items"
        ],
        "title": "PaginatedResponse[Study]"
      },
...
      "SimpleCodelistTerm": {
        "properties": {
          "codelist_uid": {
            "type": "string",
            "title": "Codelist Uid",
            "description": "Codelist UID"
          },
          "term_uid": {
            "type": "string",
            "title": "Term Uid",
            "description": "Codelist Term UID"
          }
        },
        "type": "object",
        "required": [
          "codelist_uid",
          "term_uid"
        ],
        "title": "SimpleCodelistTerm"
      },
...
      "SortByStudies": {
        "type": "string",
        "enum": [
          "uid",
          "id_prefix",
          "number"
        ],
        "title": "SortByStudies"
      },
...
      "SortOrder": {
        "type": "string",
        "enum": [
          "asc",
          "desc"
        ],
        "title": "SortOrder"
      },
      "Study": {
        "properties": {
          "uid": {
            "type": "string",
            "title": "Uid",
            "description": "Study UID"
          },
          "id": {
            "type": "string",
            "title": "Id",
            "description": "Study ID"
          },
          "id_prefix": {
            "type": "string",
            "title": "Id Prefix",
            "description": "Study ID prefix"
          },
          "number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Number",
            "description": "Study number",
            "nullable": true
          },
          "acronym": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Acronym",
            "description": "Study acronym",
            "nullable": true
          },
          "versions": {
            "items": {
              "$ref": "#/components/schemas/StudyVersion"
            },
            "type": "array",
            "title": "Versions",
            "description": "Study versions"
          },
          "data_completeness_tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Data Completeness Tags",
            "description": "List of data completeness tag names assigned to the study."
          },
          "data_suppliers": {
            "items": {
              "$ref": "#/components/schemas/StudyDataSupplier"
            },
            "type": "array",
            "title": "Data Suppliers",
            "description": "List of data suppliers of the study."
          }
        },
        "type": "object",
        "required": [
          "uid",
          "id",
          "id_prefix",
          "versions"
        ],
        "title": "Study"
      },
...
      "StudyCreateInput": {
        "properties": {
          "study_number": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^\\d{1,4}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Study Number",
            "description": "Study number, 1 to 4 digits. Together with the project ID it forms the study ID (`{project_id}-{study_number}`). Must be unique across all studies, not just within the project. At least one of `study_number` and `study_acronym` must be provided.",
            "nullable": true
          },
          "study_acronym": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Study Acronym",
            "description": "Short name of the study. Must be unique across all studies. At least one of `study_number` and `study_acronym` must be provided.",
            "nullable": true
          },
          "project_id": {
            "type": "string",
            "minLength": 1,
            "title": "Project Id",
            "description": "ID of an existing project the study belongs to. Matched exactly and case-sensitively. Valid values can be retrieved from `GET /v1/library/projects` (`id` field)."
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "Free-text description of the study",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "project_id"
        ],
        "title": "StudyCreateInput"
      },
      "StudyCreateResponse": {
        "properties": {
          "uid": {
            "type": "string",
            "title": "Uid",
            "description": "Study UID"
          },
          "id": {
            "type": "string",
            "title": "Id",
            "description": "Study ID, derived as `{project_id}-{study_number}`. The study number part is empty when the study has none, matching the `id` returned by `GET /studies`."
          },
          "acronym": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Acronym",
            "description": "Study acronym",
            "nullable": true
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "Study description",
            "nullable": true
          }
        },
        "type": "object",
        "required": [
          "uid",
          "id"
        ],
        "title": "StudyCreateResponse"
      },
...
      "StudyDataSupplier": {
        "properties": {
          "uid": {
            "type": "string",
            "title": "Uid"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "type": {
            "$ref": "#/components/schemas/SimpleCodelistTerm"
          },
          "order": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Order",
            "nullable": true
          }
        },
        "type": "object",
        "required": [
          "uid",
          "name",
          "type"
        ],
        "title": "StudyDataSupplier"
      },
...
      "StudyVersion": {
        "properties": {
          "version_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Version Status",
            "description": "Study Status",
            "nullable": true
          },
          "version_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Version Number",
            "description": "Study Version Number",
            "nullable": true
          },
          "version_started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Version Started At",
            "description": "Study Version Start Time"
          },
          "version_ended_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Version Ended At",
            "description": "Study Version End Time",
            "nullable": true
          },
          "version_author": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Version Author",
            "description": "Study Author",
            "nullable": true
          },
          "version_description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Version Description",
            "description": "Study Description",
            "nullable": true
          }
        },
        "type": "object",
        "required": [
          "version_started_at"
        ],
        "title": "StudyVersion"
      },
...
      "ValidationDetail": {
        "properties": {
          "error_code": {
            "type": "string",
            "title": "Error Code"
          },
          "field": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Field"
          },
          "msg": {
            "type": "string",
            "title": "Msg"
          },
          "ctx": {
            "title": "Ctx"
          }
        },
        "type": "object",
        "required": [
          "error_code",
          "field",
          "msg",
          "ctx"
        ],
        "title": "ValidationDetail"
      },
...
    },
    "securitySchemes": {
      "OAuth2AuthorizationCodeBearer": {
        "type": "oauth2",
        "flows": {
          "authorizationCode": {
            "scopes": {
              "api:///API.call": "Make calls to the API"
            },
            "authorizationUrl": "",
            "tokenUrl": ""
          }
        }
      },
      "BearerJwtAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "in": "header",
        "name": "Authorization",
        "description": "Access token that will be sent as `Authorization: Bearer {token}` header in all requests"
      }
    }
  },
  "servers": [
    {
      "url": "/"
    }
  ]
}